+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## main
 
 - disable libvips DoS limits on file load for libvips 8.19+
+- add lazy pretty-print and string print of main
 
 ## 9.0.17 2026/02/25